+++ src/business/data-layer/data-layer-template.js
@@ -1,0 +1,14 @@
+// this file will be replaced at build
+const dataLayer = {
+    get: function () {
+    },
+    init: function () {
+    },
+    setCookieEnabled: function () {
+    },
+    setUnsafe: function () {
+    },
+    set: function () {
+    }
+};
+export {dataLayer};