+++ src/config.js
@@ -23,7 +23,7 @@ export default {
     ],
     'storageVisitor': 'pa_vid',
     'storageUser': 'pa_user',
-    'version': '6.6.0',
+    'version': '6.7.0',
     'minHeartbeat': 5,
     'minBufferingHeartbeat': 1,
     'queueVarName': '_paq',
@@ -96,7 +96,7 @@ export default {
             },
             'no-consent': {
                 'name': 'no-consent',
-                'visitorId': 'Consent-NO',
+                'visitorId': 'no-consent',
                 'properties': {
                     'include': {
                         'visitor_privacy_consent': false,
@@ -225,6 +225,8 @@ export default {
                             'page_duration': true,
                             'page_full_name': true,
                             'page_position': true,
+                            'page_title_html': true,
+                            'pageview_id': true,
                             'previous_url': true,
                             'privacy_status': true,
                             'site': true,