+++ src/core/PianoAnalytics.js
@@ -16,9 +16,9 @@ import {Privacy} from '../business/privacy';
 import {User} from '../business/user';
 import {cloneObject} from '../utils/index';
 import {preloadTagging} from '../business/preload';
-
 import {AVInsights} from '../business/avinsights';
 import PianoAnalyticsQueue from './queue';
+import {dataLayer} from '../business/data-layer/data-layer';
 
 
 function PianoAnalytics(configuration) {
@@ -30,6 +30,19 @@ function PianoAnalytics(configuration) {
     this.setConfigurations = this.cfg.setConfigurations;
     this.getConfiguration = this.cfg.getConfiguration;
     if (BUILD_BROWSER) {
+        if (typeof window.pdl === 'undefined') {
+            window.pdl = {
+                migration: {
+                    browserId: {
+                        source: 'PA'
+                    }
+                },
+                cookies: {
+                    storageMode: 'fixed'
+                }
+            };
+        }
+
         window._pac = window._pac || {privacy: []};
         const configOverrideObject = window._pac;
         for (const config in configOverrideObject) {
@@ -37,6 +50,19 @@ function PianoAnalytics(configuration) {
                 this.setConfiguration(config, configOverrideObject[config]);
             }
         }
+        dataLayer.init({
+            cookieDefault: {
+                domain: this.getConfiguration('cookieDomain'),
+                secure: this.getConfiguration('cookieSecure'),
+                path: this.getConfiguration('cookiePath'),
+                samesite: this.getConfiguration('cookieSameSite')
+            },
+            cookies: {
+                _pcid: {
+                    expires: this.getConfiguration('storageLifetimeVisitor')
+                }
+            }
+        });
     }
     this.storage = new Storage(this);
     this.privacy = new Privacy(this);
@@ -108,20 +134,37 @@ PianoAnalytics.prototype.sendEvents = function (events, options) {
     this._queue.push(['_sendEvent', events, options]);
 };
 
+function _processCallbackIfPresent(value, cb) {
+    if (cb) {
+        cb(value);
+    }
+    return value;
+}
+
 PianoAnalytics.prototype.getVisitorId = function (callback) {
+    let forcedValue = this.getConfiguration('visitorId') || null;
     let result = null;
-    this.storage.getItem(this.getConfiguration('storageVisitor'), (function (storedValue) {
-        const value = this.getConfiguration('visitorId') || storedValue;
-        if (callback) {
-            callback(value);
-        } else {
-            result = value;
-        }
-    }).bind(this));
-    return result;
+    if (BUILD_BROWSER) {
+        result = _processCallbackIfPresent(forcedValue || dataLayer.get('browserId'), callback);
+    } else {
+        this.storage.getItem(this.getConfiguration('storageVisitor'), (function (storedValue) {
+            result = _processCallbackIfPresent(forcedValue || storedValue, callback);
+        }).bind(this));
+    }
+    if (typeof callback === 'undefined') {
+        return result;
+    }
 };
+
 PianoAnalytics.prototype.setVisitorId = function (value) {
     this.setConfiguration('visitorId', value);
+    const expirationDate = new Date();
+    expirationDate.setTime(expirationDate.getTime() + (this.getConfiguration('storageLifetimeVisitor') * 24 * 60 * 60 * 1000));
+    this.privacy.setItem(this.getConfiguration('storageVisitor'), value, expirationDate, function () {
+        if (BUILD_BROWSER) {
+            dataLayer.updateMigration();
+        }
+    });
 };
 
 PianoAnalytics.prototype.setUser = function (id, category, enableStorage) {
@@ -133,6 +176,35 @@ PianoAnalytics.prototype.getUser = function (callback) {
 PianoAnalytics.prototype.deleteUser = function () {
     this.user.deleteUser();
 };
+
+if (BUILD_BROWSER) {
+    PianoAnalytics.prototype.refresh = function () {
+        dataLayer.refresh();
+    };
+    PianoAnalytics.prototype.setContentProperty = function (name, value) {
+        const MAP_PA_DL = {
+            'content_publication_date': 'createdAt',
+            'tags_array': 'tags'
+        };
+        const temp = {};
+        if (name === 'content_publication_date' || name === 'tags_array') {
+            temp[MAP_PA_DL[name]] = value;
+        } else {
+            temp[name] = value;
+        }
+        dataLayer.set('content', temp);
+
+    };
+    PianoAnalytics.prototype.setContentProperties = function (content) {
+
+        for (const prop in content) {
+            if (Object.prototype.hasOwnProperty.call(content, prop)) {
+                this.setContentProperty(prop, content[prop]);
+            }
+        }
+
+    };
+}
 
 PianoAnalytics.prototype.PA = PianoAnalytics;