+++ src/core/steps/metadata.step.js
@@ -1,4 +1,5 @@
 import {nextStep} from './utils/index';
+import {dataLayer} from '../../business/data-layer/data-layer';
 
 function _parseLanguage(language, separators) {
     for (const separator of separators) {
@@ -10,6 +11,12 @@ function _parseLanguage(language, separators) {
     return ['', ''];
 }
 
+function _camelToSnake(string) {
+    return string.replace(/[\w]([A-Z])/g, function (m) {
+        return m[0] + '_' + m[1];
+    }).toLowerCase();
+}
+
 function metadataStep(pa, model, nextSteps) {
     model.setProperty('event_collection_platform', BUILD_BROWSER ? 'js' : 'js-browserless');
     model.setProperty('event_collection_version', model.getConfiguration('version'));
@@ -19,6 +26,23 @@ function metadataStep(pa, model, nextSteps) {
     model.setProperty('device_hour', date.getHours());
 
     if (BUILD_BROWSER) {
+        const content = dataLayer.get('content');
+        for (const prop in content) {
+            if (Object.prototype.hasOwnProperty.call(content, prop)) {
+                const MAP_PA_DL = {
+                    'createdAt': 'content_publication_date',
+                    'tags': 'tags_array'
+                };
+                if (prop === 'createdAt' || prop === 'tags') {
+                    model.setProperty(MAP_PA_DL[prop], content[prop]);
+                } else {
+                    model.setProperty(_camelToSnake(`content_${prop}`), content[prop]);
+                }
+            }
+        }
+
+        model.setProperty('has_access', dataLayer.get('userStatus'));
+        model.setProperty('pageview_id', dataLayer.get('pageViewId'));
         model.setProperty('device_screen_width', window.screen.width);
         model.setProperty('device_screen_height', window.screen.height);
         model.setProperty('device_display_width',
@@ -36,9 +60,12 @@ function metadataStep(pa, model, nextSteps) {
         model.setProperty('browser_language', languageSplitted[0]);
         model.setProperty('browser_language_local', languageSplitted[1]);
         model.setProperty('previous_url', document.referrer || '');
-        const eventUrlWithQueryString = model.getConfiguration('addEventURL') === 'true';
+        if(document.title){
+            model.setProperty('page_title_html', document.title);
+        }
+        const eventUrlWithQueryString = model.getConfiguration('addEventURL').toString() === 'true';
         if (eventUrlWithQueryString || (model.getConfiguration('addEventURL') === 'withoutQS')) {
-            model.setProperty('page_url', eventUrlWithQueryString ? window.location.href : `${window.location.protocol}//${window.location.host}${window.location.pathname}`);
+            model.setProperty('page_url', eventUrlWithQueryString ? window.location.href.split('#')[0] : `${window.location.protocol}//${window.location.host}${window.location.pathname}`);
         }
         try {
             window.navigator.userAgentData.getHighEntropyValues([